spa_projekat/requirements.txt
streamlit
gensim
pyLDAvis
pandas
openpyxl
beautifulsoup4
EbookLib
python-docx
pdfplumber
pyth
pylatexenc
matplotlib
seaborn
